% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days2lessons-package.R
\docType{package}
\name{days2lessons-package}
\alias{days2lessons}
\alias{days2lessons-package}
\title{days2lessons: Distributes Teachers Lessons On Days in a Balanced Manner}
\description{
The set of teacher/class lessons is completed with a column that allocates a day to each lesson, so that the distribution of lessons by day, by class, and by teacher is as uniform as possible. \url{https://vlad.bazon.net/}.
}
\author{
\strong{Maintainer}: Vlad Bazon \email{vlad.bazon@gmail.com}

}
\keyword{internal}
