% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days2tupl.R
\name{days2tupl}
\alias{days2tupl}
\title{Repartizarea pe zile a tuplajelor}
\usage{
days2tupl(DZ, TP)
}
\arguments{
\item{DZ}{distribuția pe zile returnată de 'mount_days()'.}

\item{TP}{setul tuplajelor, dacă există.}
}
\value{
setul tuplajelor, completat cu un câmp conținând ziua (sau zilele)
în care a fost repartizat fiecare.
}
\description{
Din distribuția pe zile a tuturor lecțiilor se deduc zilele în care
au fost repartizate tuplajele existente.
}
