% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/less2days.R
\name{less2days}
\alias{less2days}
\title{Lista lecțiilor repartizate în câte o aceeași zi}
\usage{
less2days(DZ)
}
\arguments{
\item{DZ}{distribuția pe zile a tuturor lecțiilor}
}
\value{
Listă conținând pentru fiecare zi, lecțiile acelei zile
}
\description{
Asociază fiecărei zile, setul lecțiilor repartizate în acea zi
(în format lung, 'prof/cls')
}
