% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSS}
\alias{LSS}
\title{Lecții (inclusiv cuplaje) pentru o săptămână de lucru într-o școală.}
\format{
\describe{
\item{prof}{Cod de 3 sau 6 caractere, reprezentând un profesor,
respectiv un cuplaj (doi profesori, pe grupe ale unei clase);
primele două litere abreviază disciplina pe care este încadrat
profesorul, iar cifra care urmează indexează profesorii
de pe aceeași disciplină}
\item{cls}{Clasa (două sau trei caractere) la care profesorul are de
făcut o lecție}
}
}
\usage{
LSS
}
\description{
Set numit 'LSS' de 870 lecții prof|cls (cu 66 profesori din care 4 cuplaje
și cu 32 clase). Pe săptămână, profesorii din 'LSS' au între 1 și 28 de
lecții, iar clasele între 25 și 31 de ore.
}
\keyword{datasets}
