% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount_days.R
\name{mount_days}
\alias{mount_days}
\title{Montează zilele pe lecții, echilibrat}
\usage{
mount_days(LSS, TPL = NULL, Dfh = 2)
}
\arguments{
\item{LSS}{Setul lecțiilor profesorilor și cuplajelor.}

\item{TPL}{Setul tuplajelor.}

\item{Dfh}{Diferența maximă între maximul și minimul de ore/zi la clase.}
}
\value{
O repartiție pe zile a tuturor lecțiilor, relativ echilibrată.
}
\description{
Repartizarea pe zile decurge separat: lecțiile celor neimplicați în
cuplaje (plus dacă există, cuplajele "externe"), respectiv ale cuplajelor
și membrilor acestora; în final, se adaugă repartizarea lecțiilor din
tuplaje (dacă există). Repartiția pe zile a cuplajelor și respectiv,
a tuplajelor are caracter "dinamic", repetându-se până când îmbinând cele
trei repartiții, distribuția lecțiilor la fiecare clasă este cât se poate
de echilibrată (numărul de ore/zi diferă cu cel mult 2 sau 3 ore).
}
\examples{
\donttest{
    R123 <- mount_days(LSS, TPL, Dfh = 3)
}
}
